feat: Allow building without docker support
