+++ pkg/container/docker_cli.go
@@ -1,3 +1,5 @@
+//go:build !(WITHOUT_DOCKER || !(linux || macos || windows))
+
 // This file is exact copy of https://github.com/docker/cli/blob/9ac8584acfd501c3f4da0e845e3a40ed15c85041/cli/command/container/opts.go
 // appended with license information.
 //
